requirements.txt
amqp==5.1.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.12.0
kombu==5.2.4
Markdown==3.4.1
prompt-toolkit==3.0.30
pycoingecko==2.2.0
python-decouple==3.6
pytz==2022.2
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
zipp==3.8.1
